pyproject.toml
[project]
name = "python-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.14",
    "passlib>=1.7.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pyjwt>=2.10.1",
]
